+++ assets/machines/spectrum/rom.go
@@ -7,3 +7,13 @@ import _ "embed"
 //
 //go:embed 48.rom
 var Rom48 []byte
+
+// Rom128_0 contains the binary data of the Spectrum 128K ROM 0 (128K Editor/Menu).
+//
+//go:embed 128-0.rom
+var Rom128_0 []byte
+
+// Rom128_1 contains the binary data of the Spectrum 128K ROM 1 (48K BASIC).
+//
+//go:embed 128-1.rom
+var Rom128_1 []byte